+++ io_revolt/ncp_out.py
@@ -52,7 +52,8 @@ def export_file(filepath, scene):
                 not obj.revolt.is_bbox and
                 #not obj.revolt.is_instance and
                 not obj.revolt.ignore_ncp and
-                not obj.revolt.is_mirror_plane
+                not obj.revolt.is_mirror_plane and
+                not obj.revolt.is_track_zone
             )
             if conditions:
                 objs.append(obj)